MLIBZ-1921: Swift 4 / Xcode 9 beta support

+++ Kinvey/KinveyTests/NetworkStoreTests.swift
@@ -1969,7 +1969,7 @@ class NetworkStoreTests: StoreTestCase {
         store.group(
             keys: ["name"],
             min: "age",
-            minType: Float.self,
+            minType: Double.self,
             condition: NSPredicate(format: "age > %@", NSNumber(value: 18))
         ) { (result, error) in
             XCTAssertNotNil(result)
@@ -2007,7 +2007,7 @@ class NetworkStoreTests: StoreTestCase {
         store.group(
             keys: ["name"],
             min: "age",
-            minType: Float.self,
+            minType: Double.self,
             condition: NSPredicate(format: "age > %@", NSNumber(value: 18))
         ) { (result, error) in
             XCTAssertNil(result)
@@ -2047,7 +2047,7 @@ class NetworkStoreTests: StoreTestCase {
         store.group(
             keys: ["name"],
             max: "age",
-            maxType: Float.self,
+            maxType: Double.self,
             condition: NSPredicate(format: "age > %@", NSNumber(value: 18))
         ) { (result, error) in
             XCTAssertNotNil(result)
@@ -2085,7 +2085,7 @@ class NetworkStoreTests: StoreTestCase {
         store.group(
             keys: ["name"],
             max: "age",
-            maxType: Float.self,
+            maxType: Double.self,
             condition: NSPredicate(format: "age > %@", NSNumber(value: 18))
         ) { (result, error) in
             XCTAssertNil(result)
